move tera-model-diagram off one folder
